requirements.txt
Flask==1.0.2
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
Click==7.0
cloudpickle==1.1.1
colorama==0.4.1
config==0.4.0
cycler==0.10.0
dask==1.2.2
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
entrypoints==0.3
HeapDict==1.0.0
ipynb-py-convert==0.4.4
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
llvmlite==0.28.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
msgpack==0.6.1
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numba==0.43.1
numpy==1.16.3
pandas==1.1.3
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.2
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
scikit-learn==0.21.1
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.1.0
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.5.1
zict==0.1.4
zope.interface==4.6.0
statsmodels==0.9.0
lightgbm
xlrd==1.2.0